refactor: create experience repository interface
